requirements.txt
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
langchain-chroma
langchain_core
chromadb
openai
tiktoken